feat: savings logic written
